refactor(adapter): split adapter-github.ts into PR and issue modules
Extract PR-related methods (createPr, mergePr, getPrStatus, getPrChecks,
getPrComments, getPrReviews, closePr, getPrDiff, queryPrs) into
adapter-github-pr.ts using prototype mixin pattern.

- adapter-github.ts: 200 LOC (issue CRUD, comments, labels, branch, bulk fetch)
- adapter-github-pr.ts: 137 LOC (all PR operations)
- Both files well under 300 LOC threshold
- 47/47 adapter tests pass
- Fixes Gate 4 / Constitution Rule 7 review finding

Aloop-Agent: build
Aloop-Iteration: 39
Aloop-Session: orchestrator-20260321-172932-issue-176-20260327-174859
